Bug fixes on recommendation generation in Plot Details and P & L

+++ app/src/main/java/org/grameen/fdp/kasapin/data/prefs/AppPreferencesHelper.java
@@ -60,7 +60,7 @@ public class AppPreferencesHelper implements PreferencesHelper {
 
     @Override
     public void setUserId(int userId) {
-        mPrefs.edit().putLong(PREF_KEY_USER_ID, userId).apply();
+        mPrefs.edit().putInt(PREF_KEY_USER_ID, userId).apply();
     }
 
     @Override
